added tensor_up feature
the `tensor_up()` method allows the user to easily construct the matrix
form of operators represented as PauliSums. Though there is a slight
degree of overlap with some of the existing functionality in
unitary_generators, this method provides a very easy interface to the
users of pyQuil and reference-qvm to examine operators and Hamiltonains.

In the future when the gate matrices are converted to sparse operators
we can use sparse kron in scipy to speed up construction of these
operators.  Right now there is a lot of wasted computation by
multiplying by known zeros.